More basic data structring
Added some classes to deal with an entire game managing players and chains.
Some proto-networking code is also included in the Player class.
The overview still can't process moving a chain from one player to the next;
building up work queues for each player remains as a task.
May need restructuring in the future to be thread-based.